global: refactor to feel like reading the godoc
