+++ assert.go
@@ -5,8 +5,8 @@ import (
 	"os"
 )
 
-// AssertSuccess logs the error and exists if something bad occurs
-func AssertSuccess(err error) {
+// assertSuccess logs the error and exists if something bad occurs
+func assertSuccess(err error) {
 	if err == nil {
 		return
 	}
