+++ payload.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"net"
-	"os"
+	"strings"
 )
 
 const payloadLength = 24
@@ -36,8 +36,13 @@ func NewPayload(buf []byte) (*Payload, error) {
 		return nil, errors.New("bad payload (priority should repeat)")
 	}
 
+	nicId, err := UUIDToStr(uuidbuf)
+	if err != nil {
+		return nil, err
+	}
+
 	payload := &Payload{
-		NicID:    UUIDToStr(uuidbuf),
+		NicID:    nicId,
 		Priority: buf[2],
 		ExoIP:    net.IPv4(buf[4], buf[5], buf[6], buf[7]),
 	}
@@ -45,30 +50,36 @@ func NewPayload(buf []byte) (*Payload, error) {
 	return payload, nil
 }
 
-// NetworkLoop starts the UDP server
-func (engine *Engine) NetworkLoop(listenAddress string) error {
-	ServerAddr, err := net.ResolveUDPAddr("udp", listenAddress)
-	AssertSuccess(err)
-	ServerConn, err := net.ListenUDP("udp", ServerAddr)
-	AssertSuccess(err)
-
-	buf := make([]byte, payloadLength)
-	for {
-		n, addr, err := ServerConn.ReadFromUDP(buf)
-		if err != nil {
-			Logger.Crit("network server died")
-			os.Exit(1)
-		}
-
-		if n != payloadLength {
-			Logger.Warning("bad network payload")
-		}
-
-		payload, err := NewPayload(buf)
-		if err != nil {
-			Logger.Warning("unparseable payload")
-		} else {
-			engine.UpdatePeer(*addr, payload)
-		}
+// UUIDToStr converts a UUID into a str
+func UUIDToStr(buf []byte) (string, error) {
+	if len(buf) != 16 {
+		return "", fmt.Errorf("UUID length (%d) mismatch, need 16", len(buf))
 	}
+
+	uuid := fmt.Sprintf("%x-%x-%x-%x", buf[0:4], buf[4:6], buf[6:8], buf[8:10], buf[10:16])
+
+	return uuid, nil
+}
+
+// StrToUUID convert a str into a UUID
+func StrToUUID(uuid string) ([]byte, error) {
+	if len(uuid) != 36 {
+		return nil, fmt.Errorf("UUID source length (%d) mismatch, need 36", len(uuid))
+	}
+
+	uuid = strings.Replace(strings.ToLower(uuid), "-", "", -1)
+	if len(uuid) != 32 { // 36 - 4
+		return nil, errors.New("UUID has wrong length, need 32")
+	}
+
+	ba, err := hex.DecodeString(uuid)
+	if err != nil {
+		return nil, errors.New("Invalid UUID")
+	}
+
+	if len(ba) != 16 { // 32 / 2
+		return nil, fmt.Errorf("UUID converted length (%d) mismatch, need 16", len(ba))
+	}
+
+	return ba, nil
 }
